django/todoapi/views.py
from .models import Task
from .serializers import TaskSerializer

from rest_framework import generics


from django.http import Http404
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status 

class TaskList(generics.ListCreateAPIView):
    """
    Lists and creates tasks.
    """
    queryset = Task.objects.all()
    serializer_class = TaskSerializer


class TaskDetail(generics.RetrieveUpdateDestroyAPIView):
    """
    Returns a single Task and allows updates and deletion of a Task.
    """
    queryset = Task.objects.all()
    serializer_class = TaskSerializer
    lookup_url_kwarg = 'task_id'