django/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
psycopg2

uWSGI==2.0.17.1